Google-Cloud-Functions/gcf-running-locally/requirements.txt
google-cloud-storage
google-cloud-bigquery
functions-framework==3.*
requests==2.28.1  # Include requests for the simulation script
